% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causeValidation.R
\name{causeValidation}
\alias{causeValidation}
\title{Execute the validation study}
\usage{
causeValidation(outputFolder, TAR = 60)
}
\arguments{
\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/)}

\item{TAR}{Time at risk for determining risk window}
}
\description{
Execute the validation study
}
\details{
This function will run the random forest model for classify causes of death
}
