% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createValidationTable.R
\name{createValidationTable}
\alias{createValidationTable}
\title{Execute the validation study}
\usage{
createValidationTable(outputFolder, TAR)
}
\arguments{
\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/)}

\item{TAR}{Time at risk for determining risk window}
}
\description{
Execute the validation study
}
\details{
This function will create the table for validation
}
